feat(manage_translations): finish added language. (👤: joilton silva)
